requirements.txt
fastapi[standard]
psycopg2-binary
boto3
python-jose[cryptography]
passlib[bcrypt]